Unify Scrollable and ScrollableNative in ui
Enact-DCO-1.0-Signed-off-by: YB Sung (yb.sung@lge.com)
