+++ packages/moonstone/Scrollable/useEventFocus.js
@@ -1,6 +1,6 @@
 import Spotlight from '@enact/spotlight';
 import {getRect} from '@enact/spotlight/src/utils';
-import {constants} from '@enact/ui/Scrollable/ScrollableNative';
+import {constants} from '@enact/ui/Scrollable/Scrollable';
 
 const {animationDuration, epsilon} = constants;
